+++ mtools/mplotqueries/mplotqueries.py
@@ -47,7 +47,7 @@ class MPlotQueriesTool(LogFileTool):
         self.argparser.add_argument('--ns', action='store', nargs='*', metavar='NS', help='(deprecated) use a prior mlogfilter instead. ')
         self.argparser.add_argument('--logscale', action='store_true', help='plot y-axis in logarithmic scale (default=off)')
         self.argparser.add_argument('--overlay', action='store', nargs='?', default=None, const='add', choices=['add', 'list', 'reset'], help="create combinations of several plots. Use '--overlay' to create an overlay (this will not plot anything). The first call without '--overlay' will additionally plot all existing overlays. Use '--overlay reset' to clear all overlays.")
-        self.argparser.add_argument('--type', action='store', default='duration', choices=self.plot_types.keys(), help='type of plot (default=duration).')        
+        self.argparser.add_argument('--type', action='store', default='scatter', choices=self.plot_types.keys(), help='type of plot (default=scatter with --yaxis duration).')        
         self.argparser.add_argument('--group', help="specify value to group on. Possible values depend on type of plot. All basic plot types can group on 'namespace', 'operation', 'thread', range and histogram plots can additionally group on 'log2code'.")
 
         # mutex = self.argparser.add_mutually_exclusive_group()
@@ -105,7 +105,7 @@ class MPlotQueriesTool(LogFileTool):
             self.args['group'] = 'filename'
         
         plot_instance = self.plot_types[self.args['type']](args=self.args, unknown_args=self.unknown_args)
-        
+
         for logfile in self.logfiles:
             start = None
 
@@ -114,6 +114,9 @@ class MPlotQueriesTool(LogFileTool):
                 logline = LogLine(line)
                 if not start:
                     start = logline.datetime
+
+                if logline.datetime:
+                    end = logline.datetime
 
                 if multiple_files:
                     # amend logline object with filename for group by filename
@@ -139,8 +142,6 @@ class MPlotQueriesTool(LogFileTool):
 
                     line_accepted = True
                     plot_instance.add_line(logline)
-
-            end = logline.datetime
 
             # store start and end for each logfile
             self.logfile_ranges.append( (start, end) )
@@ -305,11 +306,18 @@ class MPlotQueriesTool(LogFileTool):
     def plot(self):
         self.artists = []
         plt.figure(figsize=(12,8), dpi=100, facecolor='w', edgecolor='w')
-        axis = plt.subplot(111, )
+        axis = plt.subplot(111)
+
+        # set xlim from min to max of logfile ranges
+        xlim_min = min([dt[0] for dt in self.logfile_ranges])
+        xlim_max = max([dt[1] for dt in self.logfile_ranges])
+
+        if xlim_min == None or xlim_max == None:
+            raise SystemExit('no data to plot.')
 
         ylabel = ''
         for i, plot_inst in enumerate(sorted(self.plot_instances, key=lambda pi: pi.sort_order)):
-            self.artists.extend(plot_inst.plot(axis, i, len(self.plot_instances)))
+            self.artists.extend(plot_inst.plot(axis, i, len(self.plot_instances), (xlim_min, xlim_max) ))
             if hasattr(plot_inst, 'ylabel'):
                 ylabel = plot_inst.ylabel
             
@@ -321,13 +329,6 @@ class MPlotQueriesTool(LogFileTool):
 
         for label in axis.get_xticklabels():  # make the xtick labels pickable
             label.set_picker(True)
-
-        # set xlim from min to max of logfile ranges
-        xlim_min = min([dt[0] for dt in self.logfile_ranges])
-        xlim_max = max([dt[1] for dt in self.logfile_ranges])
-
-        if xlim_min == None or xlim_max == None:
-            raise SystemExit('no data to plot.')
             
         axis.set_xlim(date2num([xlim_min, xlim_max]))