+++ mtools/mplotqueries/plottypes/__init__.py
@@ -1,4 +1,5 @@
-from scatter_type import ScatterPlotType, DurationPlotType, NScannedNPlotType
+from scatter_type import ScatterPlotType, NScannedNPlotType
 from event_type import EventPlotType, RSStatePlotType
 from range_type import RangePlotType
 from histogram_type import HistogramPlotType
+from connchurn_type import ConnectionChurnPlotType