+++ mtools/test/__init__.py
@@ -1,0 +1,1 @@
+from test_all_import import all_tools