+++ mtools/util/cmdlinetool.py
@@ -14,7 +14,6 @@ class BaseCmdLineTool(object):
         self.argparser = argparse.ArgumentParser()
         self.argparser.add_argument('--version', action='version', version="mtools version %s" % __version__)
 
-
     def run(self, arguments=None, get_unknowns=False):
         """ Init point to execute the script. If `arguments` string is given, will evaluate the 
             arguments, else evaluates sys.argv. Any inheriting class should extend the run method 
@@ -70,5 +69,3 @@ class LogFileTool(BaseCmdLineTool):
 if __name__ == '__main__':
     tool = LogFileTool(multiple_logfiles=True, stdin_allowed=False)
     tool.run()
-    print tool.args
-